refactor: update pahts in test files
