+++ src/tests/store/metrics-store/helpers/convert-beholder-to-chart.test.js
@@ -1,5 +1,5 @@
 import '@/helpers/convert-date'
-import ConvertBeholderToChart from '@stores/metrics-store/helpers/convert-beholder-to-chart'
+import ConvertBeholderToChart from '@/modules/real-time-metrics/helpers/convert-beholder-to-chart'
 import { describe, expect, it } from 'vitest'
 import FIXTURES from './fixtures/convert-beholder-to-chart-fixtures'
