+++ src/tests/store/metrics-store/helpers/sort-object-by-key.test.js
@@ -1,4 +1,4 @@
-import SortObjectByKey from '@/stores/metrics-store/helpers/sort-object-by-key'
+import SortObjectByKey from '@/modules/real-time-metrics/helpers/sort-object-by-key'
 import { describe, expect, it } from 'vitest'
 
 const makeSut = () => {
